requirements.txt
beautifulsoup4==4.11.1
boto3==1.20.23
colorama==0.4.4
Flask==2.0.2
Flask_Cors==3.0.10
Flask_RESTful==0.3.9
GitPython==3.1.27
google_api_python_client==2.51.0
google_auth_oauthlib==0.5.1
halo==0.0.31
httplib2==0.20.2
oauth2client==4.1.3
pdfminer==20191125
pydub==0.25.1
pyfiglet==0.8.post1
pyftpdlib==1.5.6
python-dotenv==0.20.0
python_dateutil==2.8.2
requests==2.26.0
scrapetube==2.2.2
sermonaudio==6.10.3
yt_dlp==2021.9.2
pdfminer.six==20220524
gunicorn==20.1.0
